refactor: split up /order
